page fault rate
Ques. 21. Consider a scenario of demand paged memory. Page table is held in registers. It takes
8 milliseconds to service a page fault if an empty page is available or the replaced page is not
modified and 20 milliseconds if the replaced page is modified. Memory access time is 100
nanoseconds. Assume that the page to be replaced is modified 70 percent of the time. Generate a
solution to find maximum acceptable page-fault rate for access time that is not more than 200 nanoseconds.